Shift hotkeys to own directory
