+++ src/classes/hotkeys/HotkeysManager.tests.js
