+++ src/classes/hotkeys/HotkeysUtil.js
